moved the routes into the respective blueprints.
